Factor `Lazy` into a new module
