+++ crates/dom/src/element.rs
@@ -27,6 +27,7 @@ mod child_groups;
 mod child_vec;
 mod dom;
 mod event;
+mod lazy;
 
 /// Build an HTML element.
 pub struct ElementBuilderBase {
