+++ crates/dom/src/element/dom.rs
@@ -1,7 +1,6 @@
 use std::{
     cell::{RefCell, RefMut},
     fmt::{self, Display},
-    marker::PhantomData,
     rc::Rc,
 };
 
@@ -11,6 +10,7 @@ use self::{
     real::{RealElement, RealNode, RealText},
     virt::{VElement, VNode, VText},
 };
+use super::lazy::{IsThunk, Lazy};
 use crate::attribute::Attribute;
 
 mod real;
@@ -316,113 +316,8 @@ impl VNode for DomText {
 
 impl DomNode for DomText {}
 
-enum Lazy<Value, Thunk> {
-    Value(Value, PhantomData<Thunk>),
-    // TODO: feature to disable this at compile time
-    #[allow(dead_code)]
-    Thunk(Option<Thunk>),
-}
-
-impl<Value, Thunk> Lazy<Value, Thunk> {
-    #[allow(dead_code)]
-    fn new_value(x: Value) -> Self {
-        Self::Value(x, PhantomData)
-    }
-
-    fn new_thunk(x: Thunk) -> Self {
-        Self::Thunk(Some(x))
-    }
-}
-
-impl<Value, Thunk: Into<Value>> Lazy<Value, Thunk> {
-    fn value(&mut self) -> &mut Value {
-        self.value_with(Thunk::into)
-    }
-
-    fn value_with(&mut self, f: impl FnOnce(Thunk) -> Value) -> &mut Value {
-        *self = Self::Value(
-            match self {
-                Lazy::Value(value, _) => return value,
-                Lazy::Thunk(thunk) => f(thunk.take().unwrap()),
-            },
-            PhantomData,
-        );
-
-        match self {
-            Lazy::Value(value, _) => value,
-            Lazy::Thunk(_) => unreachable!(),
-        }
-    }
-
-    fn thunk(&mut self) -> &mut Thunk {
-        match self {
-            Lazy::Value(_, _) => panic!("Expected a thunk"),
-            Lazy::Thunk(thunk) => return thunk.as_mut().unwrap(),
-        }
-    }
-
-    fn map<Arg, R>(
-        &mut self,
-        arg: Arg,
-        f_virt: impl FnOnce(&mut Thunk, Arg) -> R,
-        f_real: impl FnOnce(&mut Value, Arg) -> R,
-    ) -> R {
-        match self {
-            Lazy::Value(value, _) => f_real(value, arg),
-            Lazy::Thunk(thunk) => f_virt(thunk.as_mut().unwrap(), arg),
-        }
-    }
-
-    fn map1<T: IsThunk>(
-        &mut self,
-        arg: T,
-        f_virt: impl FnOnce(&mut Thunk, T),
-        f_real: impl FnOnce(&mut Value, T),
-    ) {
-        if all_thunks([self, &arg]) {
-            f_virt(self.thunk(), arg);
-        } else {
-            f_real(self.value(), arg);
-        }
-    }
-
-    fn map2<T: IsThunk, U: IsThunk>(
-        &mut self,
-        arg0: T,
-        arg1: U,
-        f_virt: impl FnOnce(&mut Thunk, T, U),
-        f_real: impl FnOnce(&mut Value, T, U),
-    ) {
-        if all_thunks([self, &arg0, &arg1]) {
-            f_virt(self.thunk(), arg0, arg1);
-        } else {
-            f_real(self.value(), arg0, arg1);
-        }
-    }
-}
-
-impl<V, T> IsThunk for Lazy<V, T> {
-    fn is_thunk(&self) -> bool {
-        match self {
-            Lazy::Value(_, _) => false,
-            Lazy::Thunk(_) => true,
-        }
-    }
-}
-
 type LazyElement = Lazy<RealElement, VElement>;
 type LazyText = Lazy<RealText, VText>;
-
-// TODO: Typically, we'd check if `is_thunk`, `evaluate` if needed and pass the
-// arg on to a function. Each of these will borrow for Rc types. Can we find a
-// way around this? Maybe a `Borrowed` type on the `DomNode` trait?
-pub trait IsThunk {
-    fn is_thunk(&self) -> bool;
-}
-
-fn all_thunks<const COUNT: usize>(args: [&dyn IsThunk; COUNT]) -> bool {
-    args.into_iter().all(IsThunk::is_thunk)
-}
 
 impl IsThunk for DomElement {
     fn is_thunk(&self) -> bool {
@@ -433,27 +328,5 @@ impl IsThunk for DomElement {
 impl IsThunk for DomText {
     fn is_thunk(&self) -> bool {
         self.0.borrow().is_thunk()
-    }
-}
-
-impl<'a, T: IsThunk> IsThunk for &'a T {
-    fn is_thunk(&self) -> bool {
-        T::is_thunk(self)
-    }
-}
-
-impl<'a, T: IsThunk> IsThunk for &'a mut T {
-    fn is_thunk(&self) -> bool {
-        T::is_thunk(self)
-    }
-}
-
-impl<T: IsThunk> IsThunk for Option<T> {
-    fn is_thunk(&self) -> bool {
-        if let Some(x) = self {
-            x.is_thunk()
-        } else {
-            true
-        }
     }
 }
